# app/schemas/generated.py
# generated by datamodel-codegen:
#   filename:  openapi.yaml
#   timestamp: 2026-03-04T17:29:25+00:00

from __future__ import annotations

from datetime import datetime
from decimal import Decimal
from enum import Enum
from typing import Any, Dict, List, Optional
from uuid import UUID

from pydantic import BaseModel, condecimal, conint, constr, Field

from app.schemas.base import ORMBaseModel


class ProductStatus(str, Enum):
    ACTIVE = 'ACTIVE'
    INACTIVE = 'INACTIVE'
    ARCHIVED = 'ARCHIVED'


class ProductCreate(ORMBaseModel):
    name: constr(min_length=1, max_length=255)
    description: Optional[constr(max_length=4000)] = None
    price: condecimal(ge=Decimal('0.01'))
    stock: conint(ge=0)
    category: constr(min_length=1, max_length=100)
    status: ProductStatus


class ProductUpdate(ORMBaseModel):
    name: constr(min_length=1, max_length=255)
    description: Optional[constr(max_length=4000)] = None
    price: condecimal(ge=Decimal('0.01'))
    stock: conint(ge=0)
    category: constr(min_length=1, max_length=100)
    status: ProductStatus


class ProductResponse(ORMBaseModel):
    id: UUID
    name: str
    description: Optional[str] = None
    price: Decimal
    stock: int
    category: str
    status: ProductStatus
    created_at: datetime
    updated_at: datetime


class PaginatedProductResponse(ORMBaseModel):
    items: List[ProductResponse]
    totalElements: int
    page: int
    size: int


class OrderStatus(str, Enum):
    CREATED = 'CREATED'
    PAYMENT_PENDING = 'PAYMENT_PENDING'
    PAID = 'PAID'
    SHIPPED = 'SHIPPED'
    COMPLETED = 'COMPLETED'
    CANCELED = 'CANCELED'


class OrderItemCreate(ORMBaseModel):
    product_id: UUID
    quantity: conint(ge=1, le=999)


class OrderCreate(ORMBaseModel):
    items: List[OrderItemCreate] = Field(..., min_items=1, max_items=50)
    promo_code: Optional[constr(regex=r'^[A-Z0-9_]{4,20}$')] = None


class OrderUpdate(ORMBaseModel):
    items: List[OrderItemCreate] = Field(..., min_items=1, max_items=50)


class OrderItemResponse(ORMBaseModel):
    id: UUID
    product_id: UUID
    quantity: int
    price_at_order: Decimal


class OrderResponse(ORMBaseModel):
    id: UUID
    user_id: UUID
    status: OrderStatus
    promo_code_id: Optional[UUID] = None
    total_amount: Decimal
    discount_amount: Decimal
    items: List[OrderItemResponse]
    created_at: datetime
    updated_at: datetime


class DiscountType(str, Enum):
    PERCENTAGE = 'PERCENTAGE'
    FIXED_AMOUNT = 'FIXED_AMOUNT'


class PromoCodeCreate(ORMBaseModel):
    code: constr(regex=r'^[A-Z0-9_]{4,20}$')
    discount_type: DiscountType
    discount_value: condecimal(ge=Decimal('0.01'))
    min_order_amount: condecimal(ge=Decimal('0.01'))
    max_uses: conint(ge=1)
    valid_from: datetime
    valid_until: datetime


class PromoCodeResponse(ORMBaseModel):
    id: UUID
    code: str
    discount_type: DiscountType
    discount_value: Decimal
    min_order_amount: Decimal
    max_uses: int
    current_uses: int
    valid_from: datetime
    valid_until: datetime
    active: bool


class ErrorResponse(ORMBaseModel):
    error_code: str
    message: str
    details: Optional[Dict[str, Any]] = None