# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  version: 1.0.0
  description: API сервиса маркетплейса (Homework 2)
...
paths:
...
  /orders:
    post:
      tags:
        - Orders
      summary: Создание заказа
      description: 'Доступно для ролей: USER, ADMIN'
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Заказ успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Конфликт (недостаточно товара, неактивный товар и т.д.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Превышен лимит частоты создания заказа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Orders
      summary: Получение заказа по ID
      description: 'Доступно для ролей: USER, ADMIN'
      operationId: getOrderById
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '403':
          $ref: '#/components/responses/AccessDeniedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags:
        - Orders
      summary: Обновление заказа
      description: 'Доступно для ролей: USER, ADMIN'
      operationId: updateOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdate'
      responses:
        '200':
          description: Заказ успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/AccessDeniedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          description: Конфликт (недопустимый статус и т.д.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Превышен лимит частоты обновления заказа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{id}/cancel:
    post:
      tags:
        - Orders
      summary: Отмена заказа
      description: 'Доступно для ролей: USER, ADMIN'
      operationId: cancelOrder
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Заказ успешно отменен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '403':
          $ref: '#/components/responses/AccessDeniedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          description: Недопустимый переход состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/register:
    post:
      tags:
        - Auth
      summary: Регистрация пользователя
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '201':
          description: Пользователь успешно зарегистрирован
        '400':
          $ref: '#/components/responses/ValidationError'
  /auth/login:
    post:
      tags:
        - Auth
      summary: Аутентификация пользователя
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Ошибка аутентификации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Обновление access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Токен успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Невалидный refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    UserRole:
      type: string
      enum: [USER, SELLER, ADMIN]
    UserRegister:
      type: object
      required:
        - email
        - password
        - role
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          $ref: '#/components/schemas/UserRole'
    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    TokenResponse:
      type: object
      required:
        - access_token
        - refresh_token
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string

    # --- Product Schemas ---
...
    OrderStatus:
      type: string
      enum: [CREATED, PAYMENT_PENDING, PAID, SHIPPED, COMPLETED, CANCELED]
    OrderItemCreate:
      type: object
      required:
        - product_id
        - quantity
      properties:
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
          maximum: 999
    OrderCreate:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemCreate'
          minItems: 1
          maxItems: 50
        promo_code:
          type: string
          pattern: '^[A-Z0-9_]{4,20}$'
          nullable: true
    OrderUpdate:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemCreate'
          minItems: 1
          maxItems: 50
    OrderItemResponse:
      type: object
      required:
        - id
        - product_id
        - quantity
        - price_at_order
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
        price_at_order:
          type: number
          format: decimal
    OrderResponse:
      type: object
      required:
        - id
        - user_id
        - status
        - total_amount
        - discount_amount
        - items
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/OrderStatus'
        promo_code_id:
          type: string
          format: uuid
          nullable: true
        total_amount:
          type: number
          format: decimal
        discount_amount:
          type: number
          format: decimal
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # --- Promo Codes ---
...
    ErrorResponse:
      type: object
      required:
        - error_code
        - message
      properties:
        error_code:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
  responses:
    NotFoundError:
      description: Сущность не найдена
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            product:
              value:
                error_code: PRODUCT_NOT_FOUND
                message: Товар не найден по ID
            order:
              value:
                error_code: ORDER_NOT_FOUND
                message: Заказ не найден по ID
    ValidationError:
      description: Ошибка валидации входных данных
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: VALIDATION_ERROR
            message: Ошибка валидации данных
            details:
              price: "Значение должно быть больше 0"
    AccessDeniedError:
      description: Доступ запрещен
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: ACCESS_DENIED
            message: Недостаточно прав для выполнения операции
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT